Add support for filters
Filters can be declared as an interpreter which returned type is a
byte array. The "snappy" filter has been implemented in C as an
example.

In expressions, this interpreter is not used by default when accessing
the dpath, a new operator '*' used like C dereference operator can be
used to access the filtered contents. Links can be created to cast
filtered contents as a declared type and browse them like any other
dpath.

New tests should be added when filters can be implemented with inline
scripts, so that some dummy filters can be scripted and used more
easily.