bf: assert() possible with unaligned footer
Fix case where a block footer is not aligned with the body slack
fields' end would raise an assert in dbg_tracker_check_state() because
of a mismatch between tracker offset and item box offset, due to a
mismatch between offset computation of a footer (starting from max
span offset) and the registered box used size compute callback which
is of type "packed dynamic size".

The fix is to use the max span type to compute the box size for a
block that has a footer.