scope support for expression eval
Expressions can now be evaluated in a given scope, where the scope is
provided as a 'struct box'. All name resolution happens relative to
this scope, and the expression value/dpath is evaluated inside the
box's scope.

in the python API, any container object can be used for the expression
scope.