decouple filter from target item type
This is a significant refactor to allow filters and interpreters not
to be bound to their target item. A new "dpath_node" type has been
created that contains both an item and a filter (optional). It is
bound to field types and array item types. The final container item
nodes no longer contain a filter attribute.

This allows greater flexibility, e.g. it becomes possible to declare a
generic item type with its own properties, and create new named
expressions or fields with custom filters from the same type template.