remove named_exprs_check_duplicate()
this function was checking whether duplicate named expressions always
pointed to the same type, but it was buggy and the idea now is to
weaken typing so that it becomes possible to assign different types to
the same named expression, under different conditions (which allows a
form of polymorphism).

Completed test_pseudo_fs.py with duplicate named expression pointing
to diffent types (Dir or File).