python model: merge DataBlock and DataArray into common DataItem
Create a common DataItem class to represent any item from the tree,
whether a leaf interpreted item or a container type. It replaces
previous DataBlock and DataArray (and their base class DataContainer),
and is a step forward for filter generalization.

It allows the data path to be available when fetching a leaf item
directly in the returned type, removing the need for a "require_dpath"
argument. If an integer, the value can be used as an integer for
arithmetic or comparison, or simply converted with int().

Other changes:

- merge expr_evaluate_value() and expr_evaluate_dpath() into
  expr_evaluate() backends that handle value and dpath at the same
  time. This simplifies somehow the code flow.

- remove "object" argument of get_builtin_names(): until object-bound
  builtins are supported it's not needed.