cleanup: change box statement API to filter API
Because the scope can be different than the filter block in which to
get statements (attributes etc.) it's needed to separate the filter
from the scope. For this, move the existing box statement API (e.g
box_lookup_statement_internal(), box_iter_statements() etc.) into a
filter API in filter.c.