ft: experimental "file" filter support
For now the filter is called _file_, but it's meant to replace the
current "file { ... }" parsing scheme.

The idea is to allow multiple data sources to be used in the same
bitpunch schema, while allowing more flexibility even if a single
source is defined (e.g.removes the need for "file" special expression,
replaced by a named expression for a data source defined by the user).

For now it cannot replace the main block "file {}" as there's no way
yet to pass the path given by the user as a parameter.

E.g.

let hosts = _file_ {
    @path = "/etc/hosts";
} <> [] string {
    @boundary = '\n';
};