lexical scopes fixes
Add an extra "scope" field to the browse state: this is used
throughout the functions that may require a scope, instead of the
explicit "scope" argument passed along to various evaluation and
backend functions.

For convenience, a "scope" argument is still passed to the high-level
evaluation functions (expr_evaluate_internal() etc.), which if
non-NULL, will push/pop it to the browse state during its execution,
otherwise if NULL is passed, leaves the current scope unchanged.

This should fix a family of corner-cases where scoping was not correct
due to the use of the boxes and parents directly as lexical scopes.

Also, those fixes are important to support the full-fledged "file"
filter which, as a data source, does not reach the ancestor dpath, but
shall not mask the outer lexical scopes nevertheless.