fix in ast_node_get_target_filter()
for subscript operator, anchor expression may be something else than
an array or array slice, this was asserting: return NULL instead.

Fix file tests related to caching of file sources: add a
"notify_file_change()" method to call when a file is overwritten with
new contents, to clear it from the file source cache.