cleanup: better way to specify filters mapping to lists
Introduce a flag FILTER_CLASS_MAPS_LIST set when a filter represents
its underlying data as a list, as opposed to a native object or a
mapping. This allows not to rely on the ast_node_hdl type to decide
the representation to use in the python module make_python_object()
method.