refactor: convert bitpunch_schema to ast_node_hdl
A bitpunch_schema is now a filter special type (__schema__) in a
regular ast_node_hdl.

This is a step to allow integration of loaded schemas into a global
ast tree where multiple loaded schemas can coexist (e.g. one
containing the filters we're interested in, and another to load a
binary file and feed the filter), or where a schema definition can
itself import another schema from the filesystem.