implement bitpunch Board in place of DataTree
A significant change in approach: instead of considering a bitpunch
model always as a single data source linked to a single specification
object, add flexibility with a new "board" concept.

The board is an associative mapping that the user interacts with to
import and bind his different bitpunch objects (data sources and
filters, mostly) together. The goal is to bring more flexibility and
interaction, e.g. ease trial and error, while allowing combinations of
objects from multiple specs and data sources at once.

When using the API, the board is created by the user initially as an
empty workspace. Then he can add to it, specification objects, data
sources, or any other expression, and associate a name to each. One
useful expression is "some_data_source <> SomeSpecFilter", along with
setting "some_data_source" and "SomeSpecFilter" expressions on the
board separately, to obtain the same model previously available via
the DataTree model.

The env("DATASOURCE") concept has been removed in favor of the board
approach, since the user explicitly defines his data source(s) in the
board itself he can link them to spec filters explicitly too, removing
the need to declare the data source inside the spec file (and allowing
to use any sub-expression from that file as a spec for a data source,
too).

The bitpunch CLI has been adapted to the board model, where the board
is automatically initialized before the user enters commands to
interact with the board. Also added better support in the CLI for spec
objects, so that completion (at least) works better on those, but
printing a spec object (i.e. describing) is not yet supported.

The command set is now different:

Added commands

- let <name> = <expression>
- use <path/to/spec>
- import <name> <path/to/spec>
- bind <name> <path/to/datafile>
- clear
- unset <name>

Removed commands

- file <path/to/data/file> <path/to/spec>