improvement: option to defer value computation from dpath
This option of expr_evaluate_() family requests to return either a
dpath, or a value if no dpath exists, but not both. This can both
optimize and allow to access data paths even if there is an error to
compute their value (as the function would then return an error,
making it not able to return the dpath).

The python library now uses this option internally.

Also, fix error_dump_context_info(): NULL check