fix: spurious 'circular dependency' with named expr as member operator
Fix circular dependency error during compilation, when declaring
fields typed as a nested filter, where the nesting filter would create
a circular dependency would it be used as filter instead.

The fix itself consists of not setting COMPILE_TAG_SPAN_SIZE as a
dependency when compiling anchors of named expressions and member
operators.

additional improvement: display node types and compile tags in
circular dependency error message.

Add a test to check that was originally intended to test name scoping
behavior (and does), which also uncovered this issue hence checks it
as well.